+++ modules/hanna-react/src/CheckboxGroup.tsx
@@ -11,6 +11,7 @@ import Checkbox from './Checkbox';
 export type CheckboxGroupProps = TogglerGroupFieldProps & {
   layout?: 'inline';
   value?: Array<string>;
+  defaultValue?: Array<string>;
 };
 export type CheckboxGroupOption = TogglerGroupFieldOption;
 export type CheckboxGroupOptions = TogglerGroupFieldOptions;