+++ modules/hanna-react/src/RadioGroup.tsx
@@ -16,6 +16,7 @@ const Radio = (props: RadioProps) => <TogglerInput {...props} bem="Radio" type="
 export type RadioGroupProps = TogglerGroupFieldProps & {
   layout?: 'inline';
   value?: string;
+  defaultValue?: string;
 };
 export type RadioGroupOption = TogglerGroupFieldOption;
 export type RadioGroupOptions = TogglerGroupFieldOptions;
@@ -25,6 +26,7 @@ export const RadioGroup = (props: RadioGroupProps) => (
     {...props}
     bem="RadioGroup"
     modifier={props.layout}
+    isRadio
     Toggler={Radio}
   />
 );